pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "proSVD"
version = "0.0.1"
description = "Code for streaming dimension reduction with stability"
authors = [{name = "Pranjal Gupta"}]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

requires-python = ">=3.7"
keywords = ["streaming", "SVD"]
dependencies = [
    "numpy",
    "scipy",
    "matplotlib",
    "scikit-learn"
]

[tool.setuptools]
py-modules = ["proSVD", "pro_utils"]

[project.optional-dependencies]
related-methods = ["pymdptoolbox"]
